[soy-16:feat]Inventory grpc 추가하다 (#63)
* fix:inventory quantity를 integer로 변환하다

* [soy-16:feat]Inventory grpc 추가하다

* fix:테스트용 클래스 에러수정하다(삭제예정)

+++ api/inventory-api/src/main/java/api/inventory/request/AddInventoryReq.java
@@ -7,7 +7,7 @@ import jakarta.validation.constraints.PositiveOrZero;
 public record AddInventoryReq(
     @NotNull Long productId,
     @NotNull Long warehouseId,
-    @NotNull @PositiveOrZero Long totalQuantity,
+    @NotNull @PositiveOrZero Integer totalQuantity,
     @NotNull @PositiveOrZero Integer reorderLevel,
     @NotNull @Positive Integer reorderQuantity
 ) {
